requirements.txt
catenaconf==0.1.4
chromadb==0.5.5
jsonschema==4.23.0
numpy==2.2.2
openai==1.60.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
packaging==24.2
Pillow==11.1.0
pydantic==2.10.6
PyYAML==6.0.1
PyYAML==6.0.2
questionary==2.1.0
Requests==2.32.3
rich==13.9.4
sentence_transformers==3.0.1
setuptools==75.6.0
Spire.Doc==12.7.1
Spire.Pdf==10.8.1
torch==2.0.1
torchvision==0.15.2
typer==0.15.1
